Refactor WalletList and WalletListSwipeRow props
